refactor(core): extract PR/MR detection from claude events.ts (WS6)
The Claude adapter's events.ts mixed ~125 lines of PR/MR URL detection
(regexes, command matchers, parsers) into the event-dispatch pipeline.
Lift that self-contained concern into a new pr-detection.ts module and
import the handful of helpers events.ts still uses internally. External
consumers (lifecycle-manager, session, and the PR test files) now import
straight from pr-detection.js; events.ts shrinks from 744 to 618 lines.

Also drops a pre-existing unused PR_CREATE_COMMANDS import the no-op CI
never flagged. No behavior change — existing PR-detection tests cover it.
